moved log_to_console to pterodactyl_exporter.py because its only used there.
